Connected Texture work at last
Big update to integrate Continuity to allow for connected textures resource packs. I used a lot of mods/resource pack as a work around before so I removed/changed them. Blueprint, the library for Personality was incompatible with Continuity so I had to delete it and find replacements to keep the ability to sit and crawl.

- Added Statues, a mod that was in my previous modpack
- Added a new sleep mechanics with Sleep Tight. It's a test as it might not work well with Vampirism in multiplayer games.